breakdown into component
